third_party/WebKit/LayoutTests/fast/files/file-in-input-display-expected.txt

CONSOLE MESSAGE: line 50: [object DataTransfer]
